separate client definition from service instantiation
